+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  * Implemented rendering of auto-orient layers
  * Automatically embed images from dotLottie archives
  * SVG exporter improvements
+ * Automatic image packaging in dotLottie exporter
 
 ## 0.5.3 - 2020-04-21
  * Improved input options for tgsconvert